requirements.txt
streamlit
Pillow
brotli
lz4
zstandard
